// cloudreve-official-frontend/src/api/explorer.ts
export const FileType = {
  file: 0,
  folder: 1,
};

export interface FileResponse {
  type: number;
  id: string;
  name: string;
  permission?: string;
  created_at: string;
  updated_at: string;
  size: number;
  metadata?: {
    [key: string]: string;
  };
  path: string;
  shared?: boolean;
  capability?: string;
  owned?: boolean;
  folder_summary?: FolderSummary;
  extended_info?: ExtendedInfo;
  primary_entity?: string;
}

export interface FolderSummary {
  size: number;
  files: number;
  folders: number;
  completed: boolean;
  calculated_at: string;
}

export interface ExtendedInfo {
  storage_policy?: StoragePolicy;
  storage_policy_inherited: boolean;
  storage_used: number;
  shares?: Share[];
  entities?: Entity[];
  permissions?: PermissionSettingReq;
  view?: ExplorerView;
  direct_links?: DirectLink[];
}

export interface DirectLink {
  id: string;
  created_at: string;
  url: string;
  downloaded: number;
}

export interface Entity {
  id: string;
  type: number;
  created_at: string;
  storage_policy?: StoragePolicy;
  size: number;
  created_by?: User;
  encrypted_with?: EncryptionCipher;
}

export interface Share {
  id: string;
  name?: string;
  expires?: string;
  is_private?: boolean;
  share_view?: boolean;
  remain_downloads?: number;
  created_at?: string;
  price?: number;
  permission_setting?: PermissionSettingReq;
  url: string;
  visited: number;
  downloaded: number;
  expired?: boolean;
  unlocked: boolean;
  password_protected: boolean;
  source_type?: number;
  owner: User;
  permissions?: string;
  source_uri?: string;
  password?: string;
  show_readme?: boolean;
}

export enum PolicyType {
  local = "local",
  remote = "remote",
  oss = "oss",
  qiniu = "qiniu",
  onedrive = "onedrive",
  cos = "cos",
  upyun = "upyun",
  s3 = "s3",
  ks3 = "ks3",
  obs = "obs",
  load_balance = "load_balance",
}

export interface StoragePolicy {
  id: string;
  name: string;
  allowed_suffix?: string[];
  denied_suffix?: string[];
  allowed_name_regexp?: string;
  denied_name_regexp?: string;
  max_size: number;
  type: PolicyType;
  relay?: boolean;
  weight?: number;
  children?: StoragePolicy[];
  chunk_concurrency?: number;
  encryption?: boolean;
  streaming_encryption?: boolean;
}

export interface PaginationResults {
  page: number;
  page_size: number;
  total_items?: number;
  next_token?: string;
  is_cursor?: boolean;
}

export interface NavigatorProps {
  capability?: string;
  max_page_size: number;
  order_by_options: string[];
  order_direction_options: string[];
}

export interface ExplorerView {
  page_size: number;
  order?: string;
  order_direction?: string;
  view?: string;
  thumbnail?: boolean;
  columns?: ListViewColumnSetting[];
  gallery_width?: number;
}

export interface ListResponse {
  files: FileResponse[];
  pagination: PaginationResults;
  props: NavigatorProps;
  context_hint?: string;
  recursion_limit_reached?: boolean;
  mixed_type?: boolean;
  single_file_view?: boolean;
  parent?: FileResponse;
  storage_policy?: StoragePolicy;
  view?: ExplorerView;
}

export const Metadata = {
  share_redirect: "sys:shared_redirect",
  share_owner: "sys:shared_owner",
  upload_session_id: "sys:upload_session_id",
  icon_color: "customize:icon_color",
  emoji: "customize:emoji",
  live_photo: "customize:live_photo",
  tag_prefix: "tag:",
  thumbDisabled: "thumb:disabled",
  restore_uri: "sys:restore_uri",
  expected_collect_time: "sys:expected_collect_time",
  full_text_index: "sys:fulltext_index",

  // Exif
  gps_lng: "exif:longitude",
  gps_lat: "exif:latitude",
  gps_attitude: "exif:altitude",
  artist: "exif:artist",
  copy_right: "exif:copy_right",
  camera_model: "exif:camera_model",
  camera_make: "exif:camera_make",
  camera_owner_name: "exif:camera_owner_name",
  body_serial_number: "exif:body_serial_number",
  lens_make: "exif:lens_make",
  lens_model: "exif:lens_model",
  software: "exif:software",
  exposure_time: "exif:exposure_time",
  f_number: "exif:f",
  aperture_value: "exif:aperture_value",
  focal_length: "exif:focal_length",
  iso_speed_ratings: "exif:iso",
  pixel_x_dimension: "exif:x",
  pixel_y_dimension: "exif:y",
  orientation: "exif:orientation",
  taken_at: "exif:taken_at",
  flash: "exif:flash",
  image_description: "exif:image_description",
  projection_type: "exif:projection_type",
  exposure_bias_value: "exif:exposure_bias",

  // Music
  music_title: "music:title",
  music_artist: "music:artist",
  music_album: "music:album",

  // Stream
  stream_title: "stream:title",
  stream_duration: "stream:duration",
  stream_format_name: "stream:format",
  stream_format_long: "stream:formatLong",
  stream_bit_rate: "stream:bitrate",
  stream_description: "stream:description",
  stream_indexed_codec: "codec",
  stream_indexed_bitrate: "bitrate",
  stream_indexed_width: "width",
  stream_indexed_height: "height",

  // Geocoding
  street: "geocoding:street",
  locality: "geocoding:locality",
  place: "geocoding:place",
  district: "geocoding:district",
  region: "geocoding:region",
  country: "geocoding:country",
};

export const FilePermission = {
  read: 0,
  update: 1,
  create: 2,
  delete: 3,
  share: 4,
};
// ...
export const NavigatorCapability = {
  create_file: 0,
  rename_file: 1,
  set_permission: 2,
  move_to_my: 3,
  move_to_share: 4,
  move_to_trash: 5,
  upload_file: 6,
  download_file: 7,
  update_metadata: 8,
  list_children: 9,
  generate_thumb: 10,
  copy_to_my: 11,
  copy_to_share: 12,
  copy_to_trash: 13,
  delete_file: 14,
  lock_file: 15,
  soft_delete: 16,
  restore: 17,
  share: 18,
  info: 19,
  version_control: 20,
  mount: 21,
  relocate: 22,
  enter_folder: 23,
};

export interface PinFileService {
  uri: string;
  name?: string;
}

export interface MoveFileService extends MultipleUriService {
  dst: string;
  copy?: boolean;
}

export interface MetadataPatch {
  key: string;
  value?: string;
  remove?: boolean;
}

export interface PatchMetadataService extends MultipleUriService {
  patches: MetadataPatch[];
}

export interface PermissionSettingReq {
  same_group?: string;
  everyone?: string;
  other?: string;
  anonymous?: string;
  group_explicit?: {
    [key: string]: string;
  };
  user_explicit?: {
    [key: string]: string;
  };
}